feat: Migrate to configuration file (improvement/adaptability)
This commit migrates the application from using environment
variables to a configuration file for improved
maintainability and adaptability.  It addresses the
issue of the application failing with a 500 error due to
missing environment variables.  The configuration file
now holds endpoint URLs and other sensitive data.  The
`go.mod` and `go.sum` files are updated to reflect the
removal of the `godotenv` dependency.